+++ src/blocks-frontend.js
@@ -74,6 +74,38 @@ if (!Array.prototype.findIndex) {
   });
 }
 
+
+// Matches polyfill.
+if (!Element.prototype.matches) {
+	Element.prototype.matches =
+	  Element.prototype.msMatchesSelector ||
+	  Element.prototype.webkitMatchesSelector;
+}
+// element.closest() polyfill.
+if (!Element.prototype.closest) {
+	Element.prototype.closest = function(s) {
+	  var el = this;
+
+	  do {
+		if (Element.prototype.matches.call(el, s)) return el;
+		el = el.parentElement || el.parentNode;
+	  } while (el !== null && el.nodeType === 1);
+	  return null;
+	};
+}
+
+// Foreach NodeList Polyfill for IE.
+if ('NodeList' in window && !NodeList.prototype.forEach) {
+    console.info('polyfill for IE11');
+    NodeList.prototype.forEach = function (callback, thisArg) {
+      thisArg = thisArg || window;
+      for (var i = 0; i < this.length; i++) {
+        callback.call(thisArg, this[i], i, this);
+      }
+    };
+}
+
+
 // Foreach Polyfill
 // Production steps of ECMA-262, Edition 5, 15.4.4.18
 // Reference: http://es5.github.io/#x15.4.4.18