+++ src/blocks/collapsible-control/collapsible-control.js
@@ -4,6 +4,9 @@
  * A block to toggle collapsible blocks on the page
  */
 
+// Internal dependencies.
+import blockIcons from '../../components/block-icons';
+
 // WordPress dependencies.
 const { __ } = wp.i18n;
 const { registerBlockType } = wp.blocks;
@@ -24,7 +27,7 @@ registerBlockType( 'bu/collapsible-control', {
 	title: __( 'Collapsible Control' ),
 	description: __( 'Toggle Collapsible blocks on the page' ),
 	keywords: [ __( 'collapsible' ), __( 'control' ), __( 'toggle' ) ],
-	icon: <SVG viewBox="0 0 24 24" xmlns="http://www.w3.org/2000/svg"><Path fill="#c00" d="M19 7h-1V5h-4v2h-4V5H6v2H5c-1.1 0-2 .9-2 2v10h18V9c0-1.1-.9-2-2-2zm0 10H5V9h14v8z"></Path></SVG>,
+	icon: blockIcons('collapsiblecontrol'),
 	category: 'bu',
 	attributes: {
 		text: {